+++ settings.cjs
