+++ src/commands/command.js
@@ -1,0 +1,23 @@
+import { Msg } from '../message.js'
+
+export default class Command {
+  constructor (translate) {
+    this.translate = translate // object
+  }
+
+  showHelpText (message) {
+    return typeof this.helpText === 'function'
+      ? this.helpText(message)
+      : this.helpText
+  }
+
+  /**
+   * @param {Msg} _message
+   */
+  helpText (_message) { }
+
+  /**
+   * @param {Msg} _message
+   */
+  async handler (_message) { }
+}
