+++ src/settings.js
@@ -1,0 +1,16 @@
+import settings from '../settings.cjs'
+import translate from './translation.js'
+
+if (!settings.discord.token) throw new Error(translate.noDiscordToken)
+if (!settings.twitch.clientID) throw new Error(translate.noTwitchClientID)
+if (!settings.twitch.clientSecret) {
+  console.log(translate.includeTwitchClientSecret)
+  throw new Error(translate.noTwitchClientSecret)
+}
+
+if (typeof settings.cooldownTimer === 'undefined') {
+  console.log(translate.includeCooldownTimerWarning)
+  settings.cooldownTimer = 21600000
+}
+
+export default settings
